microtrack/dwi.py
"""

A module for representing diffusion weighted imaging data

"""

import os
import warnings

import numpy as np
import nibabel as ni

import descriptors as desc

class DWI(desc.ResetMixin):
    """
    A class for representing dwi data
    """
            
    def __init__(self, data, bvecs, bvals, affine=None):
        """
        Initialize a DWI object

        Parameters
        ----------
        data: str or array
            The diffusion weighted mr data provided either as a full path to a
            nifti file containing the data, or as a 4-d array.

        bvecs: str or array
            The unit vectors describing the directions of data
            acquisition. Either an 3 by n array, or a full path to a text file
            containing the 3 by n data.

        bvals: str or array
            The values of b weighting in the data acquisition. Either a 1 by n
            array, or a full path to a text file containing the values.

        affine: optional, 4 by 4 array
            The affine provided in the file can be overridden by explicitely
            setting this input variable. If this is left as None, one of two
            things will happen. If the 'data' input was a file-name, the affine
            will be read from that file. Otherwise, a warning will be issued
            and affine will default to np.eye(4).

        """
        
        # All inputs are handled essentially the same. Inputs can be either
        # strings, in which case file reads are required, or arrays, in which
        # case no file reads are needed and we assign these arrays into the
        # attributes:
        for name, val in zip(['data','bvecs','bvals'], [data, bvecs, bvals]): 
            if isinstance(val, str):
                exec("self.%s_file = '%s'"%(name, val))
            elif isinstance(val, np.ndarray):
                # This time we need to give it the name-space:
                exec("self.%s = val"%name, dict(self=self, val=val))
            else:
                e_s = "%s seems to be neither an array, "%name
                e_s += "nor a file-name\n"
                e_s += "The value provided was: %s" %val
                raise ValueError(e_s)

        # You can provide your own affine, if you want and that bypasses the
        # class method provided below as an auto-attr:
        if affine is not None:
            self.affine = np.matrix(affine)
        
    @desc.auto_attr
    def bvals(self):
        """
        If bvals were not provided as an array, read them from file
        """ 
        return np.loadtxt(self.bvals_file)
        
    @desc.auto_attr
    def bvecs(self):
        """
        If bvecs were not provided as an array, read them from file
        """ 
        return np.loadtxt(self.bvecs_file)

    @desc.auto_attr
    def data(self):
        """
        Load the data from file
        """
        return ni.load(self.data_file).get_data()

    @desc.auto_attr
    def affine(self):
        """
        Get the affine transformation of the data to world coordinates
        (relative to acpc)
        """
        if hasattr(self, 'data_file'):
            # This means that there might be an affine to read in from file.
            return np.matrix(ni.load(self.data_file).get_affine())
        else:
            w_s = "DWI data generated from array. Affine will be set to"
            w_s += " np.eye(4)"
            warnings.warn(w_s)
            return np.eye(4)

